+++ Sources/Paths/OnPlaybackStoppedAPI.swift
@@ -3,7 +3,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, you can obtain one at https://mozilla.org/MPL/2.0/.
 //
-// Copyright (c) 2022 Jellyfin & Jellyfin Contributors
+// Copyright (c) 2023 Jellyfin & Jellyfin Contributors
 //
 
 import Foundation
@@ -19,14 +19,14 @@ public extension Paths {
     struct OnPlaybackStoppedParameters {
         public var mediaSourceID: String?
         public var nextMediaType: String?
-        public var positionTicks: Int64?
+        public var positionTicks: Int?
         public var liveStreamID: String?
         public var playSessionID: String?
 
         public init(
             mediaSourceID: String? = nil,
             nextMediaType: String? = nil,
-            positionTicks: Int64? = nil,
+            positionTicks: Int? = nil,
             liveStreamID: String? = nil,
             playSessionID: String? = nil
         ) {
